pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "podcast-intel"
version = "0.1.0"
description = "Open-source podcast analysis and quality scoring framework"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Podcast Intel Team", email = "team@podcast-intel.local"}
]
keywords = ["podcast", "transcription", "nlp", "analysis", "quality-scoring"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "feedparser>=6.0.10",
    "requests>=2.31.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "python-dateutil>=2.8.0",
    "PyYAML>=6.0.0",
]

[project.optional-dependencies]
transcription = [
    "faster-whisper>=0.10.0",
    "pyannote-audio>=3.1.0",
    "torch>=2.0.0",
]
analysis = [
    "transformers>=4.35.0",
    "torch>=2.0.0",
    "numpy>=1.24.0",
    "scipy>=1.11.0",
]
search = [
    "chromadb>=0.4.0",
    "sentence-transformers>=2.2.0",
]
all = [
    "faster-whisper>=0.10.0",
    "pyannote-audio>=3.1.0",
    "torch>=2.0.0",
    "transformers>=4.35.0",
    "numpy>=1.24.0",
    "scipy>=1.11.0",
    "chromadb>=0.4.0",
    "sentence-transformers>=2.2.0",
]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.1.0",
    "mypy>=1.7.0",
]

[project.scripts]
podcast-intel = "podcast_intel.cli:main"

[project.urls]
Homepage = "https://github.com/podcast-intel/podcast-intel"
Documentation = "https://podcast-intel.readthedocs.io"
Repository = "https://github.com/podcast-intel/podcast-intel"
Issues = "https://github.com/podcast-intel/podcast-intel/issues"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
include = ["podcast_intel*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--strict-markers",
    "--strict-config",
    "--cov=podcast_intel",
    "--cov-report=term-missing:skip-covered",
]

[tool.ruff]
line-length = 100
target-version = "py310"
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
